Add case history page
