+++ apps/nextjs-website/src/lib/homepage.ts
@@ -32,7 +32,7 @@ export type HomepageProps = {
     readonly title: string;
     readonly products: readonly {
       readonly name: string;
-      readonly description: string | null;
+      readonly description?: string;
       readonly slug: string;
       readonly logo: {
         readonly name: string;
